Work around `scala/scala3#26533`
