+++ modules/cli/src/main/java/scala/cli/commands/pgp/PgpCommandsSubst.java
@@ -2,8 +2,8 @@ package scala.cli.commands.pgp;
 
 import com.oracle.svm.core.annotate.Substitute;
 import com.oracle.svm.core.annotate.TargetClass;
-import scala.cli.commands.pgp.ExternalCommand;
-import scala.cli.commands.pgp.PgpCommand;
+
+import scala.cli.commands.pgp.PgpExternalCommandsJava$;
 
 @TargetClass(className = "scala.cli.commands.pgp.PgpCommands")
 public final class PgpCommandsSubst {
@@ -13,11 +13,6 @@ public final class PgpCommandsSubst {
   }
   @Substitute
   public ExternalCommand[] allExternalCommands() {
-    return new ExternalCommand[] {
-      new PgpCreateExternal(),
-      new PgpKeyIdExternal(),
-      new PgpSignExternal(),
-      new PgpVerifyExternal()
-    };
+    return PgpExternalCommandsJava$.MODULE$.all();
   }
 }
