+++ modules/cli/src/main/scala/scala/cli/commands/config/ThrowawayPgpSecret.scala
@@ -9,6 +9,7 @@ import scala.build.EitherCps.{either, value}
 import scala.build.errors.BuildException
 import scala.build.{Logger, options as bo}
 import scala.cli.commands.pgp.PgpProxyMaker
+import scala.cli.commands.pgp.PgpProxySyntax.createKey
 import scala.cli.commands.shared.{CoursierOptions, SharedJvmOptions}
 import scala.cli.errors.PgpError
 import scala.cli.signing.shared.Secret
