+++ modules/cli/src/main/scala/scala/cli/commands/pgp/PgpExternalCommandsJava.scala
@@ -1,0 +1,11 @@
+package scala.cli.commands.pgp
+
+object PgpExternalCommandsJava {
+  def all(): Array[ExternalCommand] =
+    Array(
+      new PgpCreateExternal(),
+      new PgpKeyIdExternal(),
+      new PgpSignExternal(),
+      new PgpVerifyExternal()
+    )
+}
