+++ modules/cli/src/main/scala/scala/cli/commands/pgp/PgpProxyJava.scala
@@ -1,0 +1,5 @@
+package scala.cli.commands.pgp
+
+object PgpProxyJava {
+  def create(): PgpProxy = new PgpNativeProxy
+}
