+++ modules/cli/src/main/scala/scala/cli/commands/pgp/PgpPush.scala
@@ -4,6 +4,7 @@ import caseapp.core.RemainingArgs
 
 import scala.build.Logger
 import scala.cli.commands.ScalaCommand
+import scala.cli.commands.pgp.PgpProxySyntax.keyId
 import scala.cli.commands.util.ScalaCliSttpBackend
 
 object PgpPush extends ScalaCommand[PgpPushOptions] {
